requirements.txt
# Бот использует PyTelegramBotAPI (telebot), не python-telegram-bot
pyTelegramBotAPI>=4.14.0
python-dotenv>=1.0.0
